improve asset creation process
The old process would set every property when creating a new revision.
This would run all the setters, which isn't desired.  This changes it
to create the new revision in memory with the previous values, then
update that with the new values.  This will cause the setters to only
be called when new values are provided.